feat: support model validation
